Reorganize js bundle
