--- samples/StaticHtmlComposites/src/index.js
@@ -1,5 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-export * from './composites';
-export * from './service';
