+++ samples/StaticHtmlComposites/webpack.config.js
@@ -1,17 +1,19 @@
 const path = require('path');
 
 module.exports = {
-  entry: './src/index.js',
+  entry: {
+    chatComposite: './src/chatComposite.js',
+    callComposite: './src/callComposite.js',
+    service: './src/service.js'
+  },
   mode: 'development', // change to 'production' for optimization
   resolve: {
     extensions: ['.js']
   },
   output: {
-    filename: 'bundle.js',
+    filename: '[name].js',
     path: path.resolve(__dirname, 'dist'),
-    library: {
-      type: 'var',
-      name: 'app'
-    }
+    libraryTarget: 'umd',
+    library: '[name]'
   }
 };
